test(datetime): migrate position
